Added build script + added x64 configuration + shut up various x64 warnings.

+++ src/Average.cpp
@@ -27,10 +27,10 @@
 #include <limits>
 
 //Const
-static const size_t MIN_LEN = 3;
+static const uint32_t MIN_LEN = 3;
 static const double PI = 3.1415926535897932384626433832795028841971693993751058209749445923078164062862089986280348253421170679;
 
-Average::Average(const size_t &queue_len)
+Average::Average(const uint32_t &queue_len)
 :
 	m_queue_len(queue_len)
 {
@@ -39,7 +39,7 @@ Average::Average(const size_t &queue_len)
 		throw std::runtime_error("Filter size must be a positive and odd value!");
 	}
 	
-	for(size_t len = MIN_LEN; len <= queue_len; len += 2)
+	for(uint32_t len = MIN_LEN; len <= queue_len; len += 2)
 	{
 		initialize_weights(len);
 	}
@@ -49,7 +49,7 @@ Average::~Average()
 {
 }
 
-void Average::initialize_weights(const size_t filter_len)
+void Average::initialize_weights(const uint32_t filter_len)
 {
 	std::vector<double> weights(filter_len, 0.0);
 	double totalWeight = 0.0;
@@ -85,9 +85,9 @@ double Average::update(const double &value)
 
 	if(m_values.size() >= MIN_LEN)
 	{
-		const size_t key = ((m_values.size() % 2) == 1) ? m_values.size() : (m_values.size() - 1);
+		const uint32_t key = ((m_values.size() % 2) == 1) ? uint32_t(m_values.size()) : uint32_t(m_values.size() - 1U);
 
-		const std::unordered_map<size_t,std::vector<double>>::const_iterator weights = m_weights.find(key);
+		const std::unordered_map<uint32_t,std::vector<double>>::const_iterator weights = m_weights.find(key);
 		if(weights == m_weights.end())
 		{
 			throw std::runtime_error("Weights for current size not found!");
