+++ src/Average.h
@@ -25,20 +25,21 @@
 #include <deque>
 #include <vector>
 #include <unordered_map>
+#include <stdint.h>
 
 class Average
 {
 public:
-	Average(const size_t &queue_len);
+	Average(const uint32_t &queue_len);
 	~Average();
 
 	double update(const double &value);
 
 private:
-	void initialize_weights(const size_t filter_len);
+	void initialize_weights(const uint32_t filter_len);
 	
-	const size_t m_queue_len;
+	const uint32_t m_queue_len;
 	
 	std::deque<double> m_values;
-	std::unordered_map<size_t,std::vector<double>> m_weights;
+	std::unordered_map<uint32_t,std::vector<double>> m_weights;
 };
